==> 1. Downloading data from kaggle.ipynb <==
####Mount Google Drive.

from google.colab import drive
drive.mount('/content/drive')

current_dir has the name of the folder where the project is.
(you can change it to experiment, just make sure to change in subsequent notebooks as well :)

current_dir = 'Home Credit_Kaggle'

# set the project folder as current working directory
import os
complete_path = os.path.join('/content/drive/My Drive/Colab Notebooks/',current_dir)
os.chdir(complete_path)

# verify the current working directory
os.getcwd()

####Downloading the data from Kaggle

Upload the kaggle.json file in the project directory and then run below command, to copy it to root directory. In case .kaggle folder is not present in root directory, you might need to install kaggle api using below command.

How to get kaggle.json and what is kaggle API?
https://github.com/Kaggle/kaggle-api

# install kaggle api
!pip3 install kaggle

# if .kaggle folder is not present then create .kaggle folder
kag_path = '/root/.kaggle'
if not os.path.isdir(kag_path):
  os.mkdir(kag_path)

# copy kaggle.json file
!cp kaggle.json /root/.kaggle/kaggle.json

# make sure it was copied
!ls /root/.kaggle

Voila! Now use below commands to first create a folder named 'data' and then set it as current directory. 

data_path = os.path.join(complete_path,'data')
if not os.path.isdir(data_path):
  os.mkdir(data_path)

os.chdir(data_path)

# verify the current working directory
os.getcwd()

Now use kaggle api which will use kaggle json to download data to the current directory, which is 'data' folder.

!kaggle competitions download -c home-credit-default-risk
