==> 2. Extracting data.ipynb <==
###Mount Google Drive

from google.colab import drive
drive.mount('/content/drive')

current_dir has the name of the folder where the project is.

current_dir = 'Home Credit_Kaggle'

# set the 'data' folder in project folder as current working directory
import os
complete_path = os.path.join('/content/drive/My Drive/Colab Notebooks/',current_dir,'data')
os.chdir(complete_path)

# verify the current working directory
os.getcwd()

# view all the files in data folder
!ls

####Unzipping the files

# unzip the above zip files
!unzip application_test.csv.zip
!unzip application_train.csv.zip
!unzip bureau_balance.csv.zip
!unzip bureau.csv.zip
!unzip credit_card_balance.csv.zip
!unzip installments_payments.csv.zip
!unzip POS_CASH_balance.csv.zip
!unzip previous_application.csv.zip

# check the folder to make sure the csv files are present after extraction
!ls
